Refactor swap status from swap status card for use in swaplist
