+++ src/app/components/SwapStatusCard.jsx
@@ -1,67 +1,22 @@
 import React, { Fragment } from 'react'
 import PropTypes from 'prop-types'
 import { compose, setDisplayName, setPropTypes, defaultProps, withProps } from 'recompose'
-import { Row, Col, Card, CardBody, CardFooter, Alert, Collapse, Button } from 'reactstrap'
+import { Col, Card, CardBody, CardFooter, Alert, Collapse, Button } from 'reactstrap'
 import classNames from 'class-names'
 import config from 'Config'
 import withToggle from '../hoc/withToggle'
 
-import ChangePercent from 'Components/ChangePercent'
-import ArrowIcon from 'Components/ArrowIcon'
-import PriceArrowIcon from 'Components/PriceArrowIcon'
-import CoinIcon from 'Components/CoinIcon'
-import Units from 'Components/Units'
 import UnitsLoading from 'Components/UnitsLoading'
 import WalletLabel from 'Components/WalletLabel'
 import Spinner from 'Components/Spinner'
 import DataLayout from 'Components/DataLayout'
+import SwapStatus from 'Components/SwapStatus'
 
 const StatusFooter = ({ className, children, ...props }) => (
   <CardFooter className={classNames('font-size-xs py-2 px-3', className)} {...props}>
     {children}
   </CardFooter>
 )
-
-const priceChange = (date, asset) => {
-  const { change1, change7d, change24 } = asset
-  // falsey date => hasn't loaded yet so default to 1h change
-  date = new Date(date)
-  const hoursSinceTrade = !date ? 0 : (Date.now() - date.getTime()) / 3600000
-  const timespan = hoursSinceTrade <= 1 ? 'Last 1hr: ' : hoursSinceTrade >= 24 ? 'Last 7d: ' : 'Last 24hrs: '
-  var priceChange = hoursSinceTrade <= 1 ? change1 : hoursSinceTrade >= 24 ? change7d : change24
-  return (
-    <span>{timespan}
-      <ChangePercent>{priceChange}</ChangePercent>
-      <PriceArrowIcon 
-        className={classNames('swapChangeArrow', priceChange.isZero() ? 'd-none' : null)} 
-        size={.58} dir={priceChange < 0 ? 'down' : 'up'} 
-        color={priceChange < 0 ? 'danger' : priceChange > 0 ? 'success' : null}
-      />
-    </span>
-  )
-}
-
-const getShortStatus = (swap) => {
-  const { tx, status: { code, detailsCode, labelClass, label } } = swap
-  if (detailsCode === 'signed') {
-    return (<span className='text-success'>Signed</span>)
-  } else if (detailsCode === 'signing_unsupported') {
-    return (<span className='text-success'>Ready</span>)
-  } else if (detailsCode === 'signing') {
-    return (<span className='text-warning blink'>Awaiting signature</span>)
-  } else if (detailsCode.includes('error')) {
-    return (<span className='text-danger'>Failed</span>)
-  } else if (detailsCode === 'sending') {
-    return (<span className='text-primary'>Sending</span>)
-  } else if ((tx && tx.sent) || code === 'failed') {
-    return (<span className={labelClass}>{label}</span>)
-  } else if (detailsCode === 'unsigned') {
-    return null
-  } else if (detailsCode === 'creating_tx' || detailsCode === 'fetching_rate') {
-    return (<Spinner size='sm' inline/>)
-  }
-  return (<span>{label || code}</span>)
-}
 
 /* eslint-disable react/jsx-key */
 export default compose(
@@ -81,21 +36,20 @@ export default compose(
     expanded: null
   }),
   withToggle('expandedState'),
-  withProps(({ swap, expanded, isExpandedState, toggleExpandedState }) => ({
+  withProps(({ expanded, isExpandedState, toggleExpandedState }) => ({
     isExpanded: expanded === null ? isExpandedState : expanded,
     togglerProps: expanded === null ? { tag: Button, color: 'ultra-dark', onClick: toggleExpandedState } : {},
-    shortStatus: getShortStatus(swap),
   }))
 )(({
-  swap: {
-    orderId, sendWalletId, sendSymbol, sendAsset, sendAmount,
-    receiveWalletId, receiveSymbol, receiveAsset, receiveAmount, receiveAddress,
+  swap, showShortStatus, showDetails, isExpanded, togglerProps, expanded
+}) => {
+  const {
+    orderId, sendWalletId, sendSymbol, sendAmount,
+    receiveWalletId, receiveSymbol, receiveAmount, receiveAddress,
     error, friendlyError, rate, fee: swapFee, hasFee: hasSwapFee,
     tx: { confirmed, succeeded, hash: txHash, feeAmount: txFee, feeSymbol: txFeeSymbol },
-    status: { code, details, detailsCode }, createdAt, createdAtFormatted, initializing, isManual
-  },
-  shortStatus, showShortStatus, showDetails, isExpanded, togglerProps, expanded
-}) => {
+    status: { code, details, detailsCode }, createdAtFormatted, initializing, isManual
+  } = swap
   const isComplete = code === 'complete'
   const loadingValue = (error
     ? (<span className='text-danger'>-</span>)
@@ -106,45 +60,15 @@ export default compose(
         className={classNames('py-2 pr-3 pl-2 border-0 lh-0', { 'py-3': !receiveAmount && !sendAmount })}
         style={{ minHeight: '4rem' }}
         {...togglerProps}>
-        <Row className='gutter-0 align-items-center font-size-small text-muted'>
-          <Col xs='auto'>
-            { expanded === null ? <i style={{ transition: 'all .15s ease-in-out' }} className={classNames('fa fa-chevron-circle-down text-primary px-2 mr-2', { ['fa-rotate-180']: isExpanded })}/> : false }
-          </Col>
-          <Col>
-            <Row className='gutter-2 align-items-center text-center text-sm-left'>
-              <Col xs='12' sm='auto'><CoinIcon symbol={sendSymbol}/></Col>
-              <Col xs='12' sm>
-                <Row className='gutter-2'>
-                  <Col xs='12' className='order-sm-2 font-size-sm pt-0'>{sendAsset.name}</Col>
-                  {sendAmount ? (<Col xs='12' className='text-white'>
-                    <Units value={sendAmount} symbol={sendSymbol} prefix='-'/>
-                  </Col>) : null}
-                  {sendAsset ? (<Col xs='12' className='mt-0 pt-0 order-sm-3 font-size-xs'>{priceChange(createdAt, sendAsset)}</Col>)
-                    : null}
-                </Row>
-              </Col>
-            </Row>
-          </Col>
-          <Col xs='auto' className='text-center'>
-            <ArrowIcon inline size={1.5} dir='right' color={error ? 'danger' : 'success'}/><br/>
-            {showShortStatus && (<small className='lh-0'>{shortStatus}</small>)}
-          </Col>
-          <Col>
-            <Row className='gutter-2 align-items-center text-center text-sm-right'>
-              <Col xs='12' sm='auto' className='order-sm-2'><CoinIcon symbol={receiveSymbol}/></Col>
-              <Col xs='12' sm>
-                <Row className='gutter-2'>
-                  <Col xs='12' className='order-sm-2 font-size-sm pt-0'>{receiveAsset ? receiveAsset.name : receiveSymbol}</Col>
-                  {receiveAmount ? (<Col xs='12' className='text-white'>
-                    <UnitsLoading value={receiveAmount} symbol={receiveSymbol} error={error} prefix='+'/>
-                  </Col>) : null}
-                  {receiveAsset ? (<Col xs='12' className='mt-0 pt-0 order-sm-3 font-size-xs'>{priceChange(createdAt, receiveAsset)}</Col>)
-                    : null}
-                </Row>
-              </Col>
-            </Row>
-          </Col>
-        </Row>
+        <SwapStatus swap={swap}
+          className='align-items-center'
+          showShortStatus={showShortStatus}
+          before={expanded === null && (
+            <Col xs='auto'>
+              <i style={{ transition: 'all .15s ease-in-out' }} className={classNames('fa fa-chevron-circle-down text-primary px-2 mr-2', { ['fa-rotate-180']: isExpanded })}/>
+            </Col>
+          )}
+        />
       </CardBody>
       {error 
         ? (<StatusFooter tag={Alert} color='danger' className='m-0 text-center'>{friendlyError || error}</StatusFooter>)
